fix: add age restriction to get songs endpoint and change type of lik… (#37)
* fix: add age restriction to get songs endpoint and change type of likes column to long from int

* Update MusicStreamingService/Features/Songs/Get.cs


---------

+++ MusicStreamingService.Data/Entities/AlbumEntity.cs
@@ -15,7 +15,7 @@ public sealed record AlbumEntity : BaseUpdatableIdEntity
     /// <summary>
     /// Like count
     /// </summary>
-    public int Likes { get; } = 0;
+    public long Likes { get; } = 0;
     
     /// <summary>
     /// Id of the corresponding creator
